% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_median.R
\name{getMean}
\alias{getMean}
\title{Function to get mean value}
\usage{
getMean(x)
}
\arguments{
\item{x}{an R object. Numeric/logical vectors}
}
\value{
The mean of x
}
\description{
Function for the arithmetic mean.
}
\examples{
x <- c(1, 2)
getMean(x)
}
